[Silver III] Title: 후위 표기식2, Time: 0 ms, Memory: 1112 KB -BaekjoonHub

+++ C++17/백준/Silver/1935. 후위 표기식2/후위 표기식2.cc
@@ -4,9 +4,9 @@ double sta[30];
 double t[30];
 int main()
 {
-	double num;
+	int num;
 	int st=-1;
-	scanf("%lf", &num);
+	scanf("%d", &num);
 	scanf("%s", str);
 
 	for(int i=0;i<num;i++)
